fixed with decimal.js

+++ ui/lib/models/app/screen_models/scr4_exercise_details.dart
@@ -98,7 +98,7 @@ class Scr4CompletedWorkout {
       maxReps: json['maxReps'],
       minWeight: json['minWeight'].toDouble(),
       maxWeight: json['maxWeight'].toDouble(),
-      avgRestTimeBefore: json['avgRestTimeBefore'],
+      avgRestTimeBefore: json['avgRestTimeBefore'].toDouble(),
       completedRepsAmount: json['completedRepsAmount'],
       totalVolume: json['totalVolume'].toDouble(),
       plannedTotalVolume: json['plannedTotalVolume'].toDouble(),
